+++ rust/datafusion-raquet/src/udf/quadbin/quadbin_to_geojson.rs
@@ -86,7 +86,7 @@ fn build_geojson_array(arrays: Vec<ArrayRef>) -> RaquetDataFusionResult<Columnar
     let mut builder = StringViewBuilder::with_capacity(cells.len());
 
     for cell in cells.iter() {
-        let bbox = QuadBin::from_cell(cell.unwrap() as u64)?
+        let bbox = QuadBin::from_cell(cell.unwrap())?
             .to_tile()?
             .to_bbox_wgs84()?;
         let geojson = GeoFormats::new(bbox).to_geojson();
